Issue with loading model when running eval.py
The function get_last_save(...) only returns one value when a model directory is specified even though the return value is always expected to  be a 2-tuple. Also, if the user includes a trailing slash for a model directory this seems to cause issues. Proposed are fixes for both of these issues.