Use epoch to name final training epoch directory.
If you have 18 epochs saved and then train for 32, the previous 32nd epoch that was saved will get overwritten since FLAGS.advtrain_steps will also be set to 32.

Also, not sure if the behavior of line 88 is correct. If you resume training, that is start_epoch != 0, do you still need to add 1 to the current epoch?